+++ rasa/core/test.py
@@ -336,7 +336,7 @@ def _predict_tracker_actions(
     events = list(tracker.events)
 
     partial_tracker = DialogueStateTracker.from_events(
-        tracker.sender_id, events[:1], agent.domain.slots
+        tracker.sender_id, events[:1], agent.domain.slots, None, tracker.source_filename
     )
 
     tracker_actions = []
@@ -461,7 +461,7 @@ def log_failed_stories(failed, out_directory):
             f.write("<!-- All stories passed -->")
         else:
             for failure in failed:
-                f.write(failure.export_stories())
+                f.write(failure.export_stories(test=True))
                 f.write("\n\n")